+++ src/renderer/opengl/uniform_array.rs
@@ -90,6 +90,11 @@ impl UniformArray {
     pub fn set_image_blur_filter_coeff(&mut self, coeff: [f32; 3]) {
         self.0[48..51].copy_from_slice(&coeff);
     }
+
+    pub fn set_conic_start_angle(&mut self, angle: f32) {
+        // frag[13].x in the fragment shader; frag[12].w holds the scissor radius.
+        self.0[52] = angle;
+    }
 }
 
 impl From<&Params> for UniformArray {
@@ -114,6 +119,7 @@ impl From<&Params> for UniformArray {
         arr.set_image_blur_filter_direction(params.image_blur_filter_direction);
         arr.set_image_blur_filter_sigma(params.image_blur_filter_sigma);
         arr.set_image_blur_filter_coeff(params.image_blur_filter_coeff);
+        arr.set_conic_start_angle(params.conic_start_angle);
 
         arr
     }